Report only EulerRotation values
A lot went on in this problem. Mostly it was around trying to figure out
how effect managers work. I spent a lot of time looking at the Mouse
module, which uses Dom.onDocument to register global event listeners.
I also spent a lot of time looking at the Window module, which uses
a native module to get the screen dimensions from the effect manager.

The thing that was tripping it up was actually my decoder! Turns out not
all environments report the "absolute" property, which was the case on
iOS --- where I was testing.

It was worthwhile exploration though, as I don't think I would have
understood the way the effect manager architecture works otherwise.
I made sure to document exactly how they _do_ work for future reference.