Render CMYK flavor plot for acceleration
This is kind of a neat change, where we render a filled area for each
axis of acceleration. These are "cyan", "magenta", and "yellow", which
are colors used in printing. Combined with a css property,
mix-blend-mode, we're able to mimic the way that colors blend to create
a really cool effect. Depending on the angle you shake your device,
different colors blend together (cyan and yellow making green, all three
making black, etc).