Upgrade generated examples
There's now a small splash screen for viewing the different examples.
The examples themselves are no longer relying on elm-make to build the
HTML files that back them. Now, there are "frames" for the pages that
load them manually, so that the javascript can be run through the google
closure compiler. The `publish-examples` has been updated to run it all.