Add SlidingBuffer data structure
This encapsulates the logic of appending a new item to a fixed-size
list. The size is determined up-front and we drop items when the size
gets too big.

This is implemented really naively currently with a `List`, but what
I wanted to nail down was the type. The implementation can change
separately.